Improved Functions and starting to implement Samsung Firmware patching script
All functions got reviewed and got improved.
#Adding a path to Environment will stay there untill the user closes the program!
Need to work on Odin installation and management ith flashing... need to code automated version of flashing...
Stuck on looking for a script to patch AP_file (boot.img and ramdisk) with systemless protocoll and that script would be usefull if it would be a python script... Working on....

Added lz4 extractor tool in source code.